[#2224] Move access-related functions, classes from workshops.util
