+++ amy/extrequests/views.py
@@ -65,8 +65,6 @@ from workshops.models import (
     WorkshopRequest,
 )
 from workshops.util import (
-    OnlyForAdminsMixin,
-    admin_required,
     clean_upload_trainingrequest_manual_score,
     create_username,
     failed_to_delete,
@@ -77,6 +75,7 @@ from workshops.util import (
     update_manual_score,
     upload_trainingrequest_manual_score_csv,
 )
+from workshops.utils.access import OnlyForAdminsMixin, admin_required
 
 logger = logging.getLogger("amy.signals")
 scheduler = django_rq.get_scheduler("default")
